requirements.txt
asgiref==3.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
distlib==0.3.4
dj-database-url==0.5.0
Django==2.2.17
django-appconf==1.0.5
django-filter==21.1
django-heroku==0.3.1
django-staticfiles==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
filelock==3.6.0
ghs==0.1.3
gunicorn==20.1.0
halo==0.0.31
idna==3.3
importlib-metadata==4.11.2
jedi==0.18.1
log-symbols==0.0.14
mccabe==0.7.0
parso==0.8.3
Pillow==9.1.0
pipenv==2022.1.8
pipreqs==0.4.11
platformdirs==2.5.1
psycopg2==2.8.6
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==2.3.0
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
retry-requests==1.0.1
six==1.16.0
snowballstemmer==2.2.0
spinners==0.0.24
sqlparse==0.4.2
termcolor==1.1.0
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.2
virtualenv-clone==0.5.7
whitenoise==6.2.0
wrapt==1.14.1
yarg==0.1.9
zipp==3.7.0
